refactor(backend): adopt clean architecture structure
