+++ backend/src/__init__.py
@@ -1,0 +1,7 @@
+"""Clean Architecture package for Learning Zone backend."""
+
+from .usecases.student_service import StudentService
+from .infrastructure.persistence import StudentRepository
+from .domain.entities import StudentEntity
+
+__all__ = ["StudentService", "StudentRepository", "StudentEntity"]
