+++ backend/src/domain/repositories/__init__.py
@@ -1,0 +1,1 @@
+from .student_repository import StudentRepositoryInterface
