+++ backend/src/infrastructure/persistence/__init__.py
@@ -1,0 +1,1 @@
+from .student_repository import StudentRepository
